+++ assignment_1/src/model.h
@@ -15,11 +15,11 @@ class Model {
   public:
     Model(Env _env): env(_env) {};
 
-    Results solveTsp(const int, const std::vector<std::vector<double>>&);
+    Results solveTsp(const int, const std::vector<std::vector<int>>&);
 
   protected:
     Env env;
-    void initTsp(Prob, const int, const std::vector<std::vector<double>>&);
+    void initTsp(Prob, const int, const std::vector<std::vector<int>>&);
 
 };
